+++ browserium/standalonedriverobject/chromedriver.py
@@ -4,16 +4,16 @@
 # Purpose: The purpose of the class ChromeDriver is to check for the respective environment and download the required
 #          chrome driver for the respective environment.
 # Download the latest chromedriver object
+import os
 
+import requests
+
+from browserium.core.logger import Logger
 from browserium.core.utility import Utility
 from browserium.utility.os_type import OS_type
-from browserium.core.logger import Logger
-import requests
-import os
 
 
 class Chromedriver(object):
-
     def __init__(self):
         self.ut = Utility()
         self.ot = OS_type()
@@ -29,12 +29,7 @@ class Chromedriver(object):
         latest_release = config_parser.get('ChromeDriver', 'latest_browser_driver')
         arch = config_parser.get('ChromeDriver', 'arch_type')
 
-        config_dict = {
-            'driver_type':driver_type,
-            'api_url':api_url,
-            'latest_release':latest_release,
-            'arch':arch
-        }
+        config_dict = {'driver_type': driver_type, 'api_url': api_url, 'latest_release': latest_release, 'arch': arch}
 
         return config_dict
 
@@ -43,15 +38,16 @@ class Chromedriver(object):
     def url_builder(self, os_extension):
         data = self.chromedriver_objects()
         LATEST_RELEASE = requests.get(data['latest_release'])
-        url_builder = data['api_url']+LATEST_RELEASE.text+'/'+data['driver_type']+os_extension+data['arch']+'.zip'
+        url_builder = data['api_url'] + LATEST_RELEASE.text + '/' + data['driver_type'] + os_extension + data['arch'
+                                                                                                              ] + '.zip'
         return url_builder
 
     # Download the required chromedriver binary based on the operating system type
     def evaluate_on_environment(self, os_name, arch_type):
         dir_path = self.ut.get_driver_path("/dependencies/dir_chromedriver")
         if os_name == 'macos' and arch_type == '64':
-            self.ut.log_message("INFO", "Environment: "+os_name)
-            self.ut.log_message("INFO", "Architecture Type: "+arch_type)
+            self.ut.log_message("INFO", "Environment: " + os_name)
+            self.ut.log_message("INFO", "Architecture Type: " + arch_type)
             url_builder_mac = self.url_builder('_mac')
             self.ut.log_message("INFO", "Downloading the required binary for chromedriver")
             self.ut.driver_downloader(url_builder_mac, dir_path)
@@ -59,8 +55,8 @@ class Chromedriver(object):
             self.ut.unzip_file('dir_chromedriver/')
             self.ut.log_message("INFO", "Unarchiving contents completed")
         if os_name == 'linux' and arch_type == '64':
-            self.ut.log_message("INFO", "Environment: "+os_name)
-            self.ut.log_message("INFO", "Architecture Type: "+arch_type)
+            self.ut.log_message("INFO", "Environment: " + os_name)
+            self.ut.log_message("INFO", "Architecture Type: " + arch_type)
             url_builder_linux = self.url_builder('_linux')
             self.ut.log_message("INFO", "Downloading the required binary for chromedriver")
             self.ut.driver_downloader(url_builder_linux, dir_path)
@@ -73,7 +69,10 @@ class Chromedriver(object):
     def download_driver(self):
         dir_path = self.ut.get_driver_path("/dependencies/dir_chromedriver")
         if os.path.exists(dir_path):
-            self.ut.log_message("INFO", "chrome driver is already present. To update chromedriver please run `browserium update --driver=chromedriver`")
+            self.ut.log_message(
+                "INFO",
+                "chrome driver is already present. To update chromedriver please run `browserium update --driver=chromedriver`"
+            )
         else:
             os.makedirs(dir_path)
             os_name = self.ot.os_name()
@@ -83,9 +82,9 @@ class Chromedriver(object):
     # Update the required chromedriver based on the operating system type
     def update_driver(self):
         self.ut.check_directory_content("/dependencies/dir_chromedriver/chromedriver")
-        self.ut.log_message("INFO","Deleting directory contents")
+        self.ut.log_message("INFO", "Deleting directory contents")
         self.ut.delete_dir_contents('dir_chromedriver/')
         os_name = self.ot.os_name()
         arch_type = str(self.ot.os_architecture())
         self.evaluate_on_environment(os_name, arch_type)
-        self.ut.log_message("INFO","chromedriver updated")
+        self.ut.log_message("INFO", "chromedriver updated")
