+++ browserium/standalonedriverobject/geckodriver.py
@@ -4,15 +4,14 @@
 # Purpose: The purpose of the class Geckodriver is to check for the respective environment and download the required
 #          chrome driver for the respective environment.
 # Download the latest gecko driver object
+import os
 
+from browserium.core.logger import Logger
 from browserium.core.utility import Utility
 from browserium.utility.os_type import OS_type
-from browserium.core.logger import Logger
-import os
 
 
 class Geckodriver(object):
-
     def __init__(self):
         self.ut = Utility()
         self.ot = OS_type()
@@ -74,7 +73,8 @@ class Geckodriver(object):
         if os.path.exists(dir_path):
             self.log.log_info(
                 "gecko driver is already present. To update gecko driver please run `browserium update "
-                "--driver=geckodriver`")
+                "--driver=geckodriver`"
+            )
         else:
             os.makedirs(dir_path)
             os_name = self.ot.os_name()
